Move platform specific AsyncIO implementations to separate files
